requirements.txt
kaleido
numpy
pandas
customtkinter
plotly
pillow
scipy
pytest
